requirements.txt
openpyxl
colorama
requests
python-dotenv
google-generativeai
